refactor(Pit Connect): continue clean up
