+++ packages/blockchain-wallet-v4-frontend/src/scenes/Home/model.js
@@ -2,85 +2,21 @@ import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import React, { useEffect, useState } from 'react'
 import { FormattedMessage } from 'react-intl'
-import styled, { keyframes } from 'styled-components'
+import { Button } from 'blockchain-info-components'
 
-import { Button, Icon, Image, Text } from 'blockchain-info-components'
+import {
+  CloseTourIcon,
+  StepImg,
+  StepIcon,
+  StepTitle,
+  StepContent,
+  TooltipBody,
+  TooltipContent,
+  TooltipFooter
+} from 'components/Tour'
 import { actions, model } from 'data'
 
 const { GENERAL_EVENTS } = model.analytics
-
-const Scale = () => {
-  return keyframes`
-    0% {
-      opacity: 0;
-      transform: scale(0);
-    }
-    100% {
-      transform: scale(1);
-      opacity: 1;
-    }
-  `
-}
-
-const TooltipBody = styled.div`
-  position: relative;
-  width: 100%;
-  max-width: 256px;
-  background-color: ${props => props.theme['white']};
-  border-radius: 8px;
-  box-shadow: 0 4px 32px rgba(5, 24, 61, 0.4);
-  padding: 32px;
-  animation: ${Scale} 0.3s ease-in-out;
-
-  > span:first-child {
-    position: absolute;
-    top: 24px;
-    right: 24px;
-
-    &:hover {
-      cursor: pointer;
-    }
-  }
-`
-const TooltipContent = styled.div`
-  color: ${props => props.theme['white']};
-  margin-bottom: 24px;
-  text-align: center;
-`
-const TooltipFooter = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: ${props =>
-    props.isLastStep ? 'flex-end' : 'space-between'};
-  align-content: center;
-  align-items: center;
-  color: ${props => props.theme['white']};
-`
-
-const StepImg = styled(Image)`
-  margin-bottom: 32px;
-`
-
-const StepTitle = styled(Text)`
-  font-size: 20px;
-  text-align: center;
-  margin-bottom: 8px;
-  line-height: 24px;
-`
-
-const StepContent = styled(Text)`
-  line-height: 24px;
-`
-
-const CloseTourIcon = styled(Icon)`
-  &:hover {
-    color: ${({ theme }) => theme['grey600']};
-  }
-
-  &:active {
-    color: ${({ theme }) => theme['grey800']};
-  }
-`
 
 const TourTooltipComponent = ({
   analyticsActions,
@@ -265,6 +201,28 @@ export const TOUR_STEPS = [
     ),
     placement: 'right',
     disableBeacon: true
+  },
+  {
+    target: '.the-pit-tooltip',
+    content: (
+      <>
+        <StepIcon name='the-pit' size='56px' color='pitBlue' />
+        <StepTitle size='20px' weight={600}>
+          <FormattedMessage
+            id='the.pit.tooltip.title'
+            defaultMessage='Trade in The PIT.'
+          />
+        </StepTitle>
+        <StepContent color='grey600' size='14px' weight={500}>
+          <FormattedMessage
+            id='wallettour.pit.tooltip.content'
+            defaultMessage="Now that you have a Wallet, link and exchange over 26 pairs in The PIT - Blockchain's own lightning fast crypto exchange."
+          />
+        </StepContent>
+      </>
+    ),
+    disableBeacon: true,
+    placement: 'right'
   }
 ]
